backend/requirements.txt
Flask==3.1.1
Flask-JWT-Extended==4.7.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
paho-mqtt==2.1.0
psycopg[binary]==3.2.9
python-dotenv==1.0.1
pytest==8.4.1
psycopg2